Split resources into multiple modules
