+++ rest-service/manager_rest/rest/resources/__init__.py
@@ -1,0 +1,46 @@
+#########
+# Copyright (c) 2017 GigaSpaces Technologies Ltd. All rights reserved
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+#  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  * See the License for the specific language governing permissions and
+#  * limitations under the License.
+#
+
+from .blueprints import ( # noqa
+    Blueprints,
+    BlueprintsId,
+    BlueprintsIdArchive,
+)
+from .deployments import ( # noqa
+    Deployments,
+    DeploymentsId,
+    DeploymentsIdOutputs,
+    DeploymentModifications,
+    DeploymentModificationsId,
+    DeploymentModificationsIdFinish,
+    DeploymentModificationsIdRollback,
+)
+from .executions import ( # noqa
+    Executions,
+    ExecutionsId,
+)
+from .evaluate_functions import EvaluateFunctions  # noqa
+from .events import Events  # noqa
+from .nodes import ( # noqa
+    Nodes,
+    NodeInstances,
+    NodeInstancesId,
+)
+from .provider_context import ProviderContext  # noqa
+from .search import Search  # noqa
+from .status import Status  # noqa
+from .tokens import Tokens  # noqa
+from .version import Version  # noqa
