+++ rest-service/manager_rest/rest/resources/search.py
@@ -1,0 +1,52 @@
+#########
+# Copyright (c) 2017 GigaSpaces Technologies Ltd. All rights reserved
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+#  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  * See the License for the specific language governing permissions and
+#  * limitations under the License.
+#
+
+from flask_restful_swagger import swagger
+
+from manager_rest.rest.rest_decorators import (
+    exceptions_handled,
+    insecure_rest_method,
+)
+from manager_rest.rest.rest_utils import get_json_and_verify_params
+from manager_rest.security import SecuredResource
+from manager_rest.storage import ManagerElasticsearch
+
+
+class Search(SecuredResource):
+
+    @swagger.operation(
+        nickname='search',
+        notes='Returns results from the storage for the provided '
+              'ElasticSearch query. The response format is as ElasticSearch '
+              'response format.',
+        parameters=[{'name': 'body',
+                     'description': 'ElasticSearch query.',
+                     'required': True,
+                     'allowMultiple': False,
+                     'dataType': 'string',
+                     'paramType': 'body'}],
+        consumes=['application/json']
+    )
+    @exceptions_handled
+    @insecure_rest_method
+    def post(self, **kwargs):
+        """
+        Search using an Elasticsearch query
+        """
+        request_dict = get_json_and_verify_params()
+        return ManagerElasticsearch.search(
+            index='cloudify_storage',
+            body=request_dict)
